+++ src/ZPublisher/tests/test_WSGIPublisher.py
@@ -444,6 +444,73 @@ class TestPublishModule(ZopeTestCase):
         app_iter = self._callFUT(environ, start_response, _publish)
         self.assertTrue(app_iter is body)
 
+    def test_stream_file_wrapper(self):
+        from ZPublisher.Iterators import IStreamIterator
+        from zope.interface import implementer
+        from ZPublisher.HTTPResponse import WSGIResponse
+
+        @implementer(IStreamIterator)
+        class TestStreamIterator(object):
+            data = "hello" * 20
+
+            def __len__(self):
+                return len(self.data)
+
+        class Wrapper(object):
+            def __init__(self, file):
+                self.file = file
+
+        _response = WSGIResponse()
+        _response.setHeader('Content-Type', 'text/plain')
+        body = _response.body = TestStreamIterator()
+        environ = self._makeEnviron(**{'wsgi.file_wrapper': Wrapper})
+        start_response = DummyCallable()
+        _publish = DummyCallable()
+        _publish._result = _response
+        app_iter = self._callFUT(environ, start_response, _publish)
+        self.assertTrue(app_iter.file is body)
+        self.assertTrue(isinstance(app_iter, Wrapper))
+        self.assertEqual(
+            int(_response.headers['content-length']), len(body))
+        self.assertTrue(
+            _response.headers['content-type'].startswith('text/plain'))
+        self.assertEqual(_response.status, 200)
+
+    def test_unboundstream_file_wrapper(self):
+        from ZPublisher.Iterators import IUnboundStreamIterator
+        from zope.interface import implementer
+        from ZPublisher.HTTPResponse import WSGIResponse
+
+        @implementer(IUnboundStreamIterator)
+        class TestUnboundStreamIterator(object):
+            data = "hello"
+
+            def __len__(self):
+                return len(self.data)
+
+        class Wrapper(object):
+            def __init__(self, file):
+                self.file = file
+
+        _response = WSGIResponse()
+        _response.setStatus(200)
+        # UnboundStream needs Content-Length header
+        _response.setHeader('Content-Length', '5')
+        _response.setHeader('Content-Type', 'text/plain')
+        body = _response.body = TestUnboundStreamIterator()
+        environ = self._makeEnviron(**{'wsgi.file_wrapper': Wrapper})
+        start_response = DummyCallable()
+        _publish = DummyCallable()
+        _publish._result = _response
+        app_iter = self._callFUT(environ, start_response, _publish)
+        self.assertTrue(app_iter.file is body)
+        self.assertTrue(isinstance(app_iter, Wrapper))
+        self.assertEqual(
+            int(_response.headers['content-length']), len(body))
+        self.assertTrue(
+            _response.headers['content-type'].startswith('text/plain'))
+        self.assertEqual(_response.status, 200)
+
     def test_request_closed(self):
         environ = self._makeEnviron()
         start_response = DummyCallable()